+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # v 2.11.0 (?)
 Changes in this release:
 - Remove io support to be compatible with ISO8
+- Add strict mode in `get_resource` method to  the `project` fixture and `Result` object (by default set to `False`). 
+
+## Updating
+- All uses of the `get_resource` method should be replaced with the `get_one_resource` method. It has improved type matching and raises an exception when multiple matches are present.
 
 # v 2.10.0 (2024-07-05)
 Changes in this release:
 - Add support for custom `agent_map` enabling testing or remote io scenarios.
 - Fix resource serialization (#437)
 - Fix new agent api in inmanta.agent.executor.AgentInstance (ISO8)
+- Add `get_one_resource` method to the `project` fixture and `Result` object.
 
 # v 2.9.0 (2023-11-29)
 Changes in this release: